pyproject.toml
[project]
name = "open-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "levenshtein>=0.27.1",
    "lxml>=5.4.0",
    "peft>=0.15.2",
    "pillow>=11.2.1",
    "torch>=2.7.0",
    "transformers>=4.51.3",
]
